fix error p specifier

+++ src/libft1.c
@@ -37,7 +37,7 @@ void ft_putstr(char *s)
 	// }
 }
 
-void ft_putnbr(int nb)
+void ft_putnbr(long long nb)
 {
 	int i;
